+++ source/main/network/Network.cpp
@@ -678,7 +678,7 @@ bool ConnectThread()
     if (!m_password.IsEmpty())
     {
         RoR::CSHA1 sha1;
-        sha1.UpdateHash((uint8_t*)m_password.GetBuffer(), m_password.GetLength());
+        sha1.UpdateHash((uint8_t*)m_password.GetBuffer(), static_cast<uint32_t>(m_password.GetLength()));
         sha1.Final();
         sha1.ReportHash(sha1pwresult, RoR::CSHA1::REPORT_HEX_SHORT);
     }
